Create util class CSVFormat
